+++ rubicon/objc/__init__.py
@@ -12,10 +12,12 @@ from . import runtime  # noqa: F401
 from . import collections  # noqa: F401
 
 from .runtime import (  # noqa: F401
-    IMP, SEL, Block, Class, Ivar, Method, NSArray, NSDictionary, NSMutableArray, NSMutableDictionary, NSObject,
-    NSObjectProtocol, ObjCBlock, ObjCClass, ObjCInstance, ObjCMetaClass, ObjCProtocol, at, get_ivar, ns_from_py,
-    objc_classmethod, objc_const, objc_id, objc_ivar, objc_method, objc_property, objc_property_t, objc_rawmethod,
-    py_from_ns, send_message, send_super, set_ivar,
+    IMP, SEL, Class, Ivar, Method, get_ivar, objc_id, objc_property_t, send_message, send_super, set_ivar,
+)
+from .api import (  # noqa: F401
+    Block, NSArray, NSDictionary, NSMutableArray, NSMutableDictionary, NSObject, NSObjectProtocol, ObjCBlock,
+    ObjCClass, ObjCInstance, ObjCMetaClass, ObjCProtocol, at, ns_from_py, objc_classmethod, objc_const, objc_ivar,
+    objc_method, objc_property, objc_rawmethod, py_from_ns,
 )
 from .types import (  # noqa: F401
     CFIndex, CFRange, CGFloat, CGGlyph, CGPoint, CGPointMake, CGRect,
