+++ rubicon/objc/collections.py
@@ -1,9 +1,10 @@
 import operator
 
 from .types import NSUInteger, NSNotFound, NSRange, unichar
-from .runtime import (
+from .runtime import objc_id, send_message
+from .api import (
     NSArray, NSDictionary, NSMutableArray, NSMutableDictionary, NSString, ObjCInstance, for_objcclass, ns_from_py,
-    objc_id, py_from_ns, send_message,
+    py_from_ns,
 )
